Implement the Node and Tree classes
The Node class represents a single node in the binary search tree which
consists of a data value, a pointer to a left child, and a pointer to a
right child.

The Tree class contains all the methods necessary for a binary search
tree to function. The only property a Tree class contains is the root
node of the binary search tree. A buildTree function has also been
implemented in order to initialize a balanced binary search tree when
an array is given to a constructor of a Tree class.